Fix in TOpt::parseOptions()

+++ Options/Opt.cpp
@@ -120,7 +120,7 @@ bool TOpt::parseOptions(TContainer< SPtr<TOpt> >& options,
                         const char* buf,
                         size_t len,
                         TMsg* parent,
-                        uint8_t placeId /*= 0*/, // 5 (option 5) or (message 5)
+                        uint16_t placeId /*= 0*/, // 5 (option 5) or (message 5)
                         std::string place /*= "option"*/ // "option" or "message"
                         ) {
